# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minha API",
        "description": "Documentação da API",
        "contact": {
            "email": "suporte@minhaapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transaction/getall": {
            "get": {
                "tags": [
                    "Transacoes"
                ],
                "summary": "Retrieve all transactions for the authenticated user",
                "operationId": "1c17c71bdd5caf9b66008bb2c73e50ce",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transacao"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not Authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User not Authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transactions not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Transactions not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error message"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transaction/create": {
            "post": {
                "tags": [
                    "Transacao"
                ],
                "summary": "Create a new transaction",
                "operationId": "70ddc4839b8e55cd3e4b43b4612627e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "t_tipo",
                                    "value",
                                    "categoria_id"
                                ],
                                "properties": {
                                    "t_tipo": {
                                        "type": "string",
                                        "example": "expense"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction created successfully"
                                        },
                                        "transaction": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed Request Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Conflict"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The provided customer ID does not exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Conflict"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The provided customer ID does not exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create transaction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to create transaction"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Transacao": {
                "title": "Transacao",
                "description": "Modelo de Transacao",
                "properties": {
                    "value": {
                        "description": "Valor da transação",
                        "type": "number",
                        "format": "float"
                    },
                    "user_id": {
                        "description": "ID do usuário",
                        "type": "integer"
                    },
                    "categoria_id": {
                        "description": "ID da categoria",
                        "type": "integer"
                    },
                    "t_tipo": {
                        "description": "Tipo da transação",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}